backend/requirements.txt
fastapi
uvicorn
websockets
openai
elevenlabs
python-multipart
pydantic
python-dotenv
aiofiles